Refactor reporters and add KB URL to reports (#275)
* refactor reporters and add kburl

* rename json reporter file to align with other reporters
